requirements.txt
fastapi==0.115.5
openai==1.55.3
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
streamlit==1.40.2
uvicorn==0.32.1